mcp/zana-episodic/pyproject.toml
[project]
name = "zana-episodic-mcp"
version = "0.1.0"
description = "ZANA MCP Episodic Memory Server"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
zana-episodic = "server:main"

[tool.uv]
package = false
